management/requirement.txt
amqp==5.0.2
appdirs==1.4.4
arabic-reshaper==2.1.3
asgiref==3.2.10
asn1crypto==1.4.0
attrs==20.3.0
backcall==0.2.0
bcrypt==3.2.0
beautifulsoup4==4.9.3
billiard==3.6.3.0
black==20.8b1
boto3==1.16.58
botocore==1.19.58
bson==0.5.10
celery==5.0.5
certifi==2020.6.20
cffi==1.14.5
cfgv==3.2.0
chardet==3.0.4
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.1.6
crypto==1.4.1
cryptography==3.4.6
decorator==4.4.2
distlib==0.3.1
Django==3.1.2
django-celery-beat==2.2.0
django-core==1.4.1
django-cors-headers==3.5.0
django-crispy-forms==1.11.2
django-defender==0.8.0
django-extensions==3.1.0
django-filter==2.4.0
django-ipware==3.0.2
django-oauth-toolkit==1.3.3
django-oauth2==3.2
django-ratelimit==3.0.1
django-rename-app==0.1.3
django-reset-migrations==0.4.0
django-s3-storage==0.13.4
django-simple-history==2.12.0
django-smtp-ssl==1.0
django-storages==1.11.1
django-structlog==2.1.0
django-tempus-dominus==5.1.2.13
django-timezone-field==4.1.1
django-tracking==0.2.7
django-truncate==0.1
djangorestframework==3.12.1
djangorestframework-oauth==1.1.0
docopt==0.6.2
endesive==2.0.3
et-xmlfile==1.1.0
filelock==3.0.12
future==0.18.2
html2text==2020.1.16
html5lib==1.1
identify==1.5.10
idna==2.10
importlib-metadata==3.1.0
importlib-resources==3.3.0
ipdb==0.13.4
ipython==7.16.1
ipython-genutils==0.2.0
jedi==0.17.2
jmespath==0.10.0
kombu==5.0.2
listener==2.0.0a1
lxml==4.6.2
mypy-extensions==0.4.3
Naked==0.1.31
nodeenv==1.5.0
num2words==0.5.10
numpy==1.19.4
oauthlib==3.1.0
openpyxl==3.0.7
oscrypto==1.2.1
pandas==1.1.5
paramiko==2.7.2
parso==0.7.1
pathspec==0.8.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.0.1
pre-commit==2.9.2
prompt-toolkit==3.0.8
psycopg2-binary==2.8.6
ptyprocess==0.6.0
pycparser==2.20
pycrypto==2.6.1
pycryptodome==3.10.1
pycryptodomex==3.10.1
Pygments==2.7.1
#pygraphviz==1.6
PyKCS11==1.5.10
PyNaCl==1.4.0
pyOpenSSL==20.0.1
PyPDF2==1.26.0
python-bidi==0.4.2
python-crontab==2.5.1
python-dateutil==2.8.1
pytube==10.5.3
pytz==2020.1
PyYAML==5.3.1
redis==3.5.3
regex==2021.4.4
reportlab==3.5.67
requests==2.24.0
s3transfer==0.3.4
shellescape==3.8.1
shortuuid==1.0.1
six==1.15.0
soupsieve==2.2.1
sqlparse==0.4.1
structlog==20.1.0
toml==0.10.2
traitlets==4.3.3
typed-ast==1.4.3
typing-extensions==3.7.4.3
urllib3==1.25.11
vine==5.0.0
virtualenv==20.2.1
wcwidth==0.2.5
webencodings==0.5.1
#xhtml2pdf==0.2.5
xlrd==1.2.0
xlwt==1.3.0
zipp==3.4.0
django-mathfilters==1.0.0
